Changed enabled and disabled methods
The previous methods cause a StackOverflow Exception. Since these
methods are called in hardware methods that already call disable() or
enable(), the listener does not have to perform these functions or do
any notifying. I'm guessing these methods will be important for error
messages and features that are part of the next assignments but for now
I don't think they're necessary for handling coin insertion and button
presses.